idk from stribor

+++ oengine/ecs/Entity.odin
@@ -10,7 +10,7 @@ ECS_FIXED :: 2
 
 MAX_CMPNTS :: 64
 
-EntFlags :: distinct bit_set[u32(0)..=127]
+EntFlags :: distinct bit_set[0..=127]
 
 FrustumType :: enum {
     INTERNAL = 0,
